rename delivery-address-map.jsx in helpers
